wip
Calendar.re is more or less what I wanted to start with, but I couldn't
get it to compile.

Calendar2.re is Component2.re, but with some modifications to slowly
make it into Calendar.re.

At this point, I'm bailing on ReasonML because the docs really kind of
suck. The docs are a nice intro, but don't cover things like "what does
this function do" or "what functions are available on this built-in
type".