wip - I need to adjust the "finished" count if the date has changed
The elm script always gets the current date and overwrites the one from
local storage. If it does, and today is different than the last time I
loaded this, I want to adjust the "days finished" to match what it would
have been with the old calendar.

This is part of the way there. It adds an 'adjustFinished' func that has
the signature I want, and sets up a WIP version of it. On the JS side, I
force the date to be a week ago on load so that I can test this.